GenreWordReplacerFactoryを除去して、単一モジュールに統合
